feat: moved files.go into seperate package
